refact(p2p): move things and call them by the right name
