+++ usecases/get_node.py
@@ -1,12 +1,12 @@
 from typing import Union
 
-from gateways.cache.cache_gateway import CacheGateway
+from gateways.node_gateway import NodeGateway
 
 
 class GetNode:
 
-    def __init__(self, cache_gateway: Union[CacheGateway, None] = None) -> None:
-        self.cache_gateway = cache_gateway or CacheGateway()
+    def __init__(self, cache_gateway: Union[NodeGateway, None] = None) -> None:
+        self.cache_gateway = cache_gateway or NodeGateway()
 
     def get(self, id: str) -> Union[dict, None]:
         network = self.cache_gateway.get_node(id)
